Fix problem with reloading language.yml file
